Задача 1: исправлены неверные типы указателей

+++ HW-7/src/prob-1.c
@@ -7,14 +7,14 @@
 
 int main() {
     float nums[5];
-    for (int *it = nums, *end = nums + 5; it != end; ++it)
+    for (float *it = nums, *end = nums + 5; it != end; ++it)
         if (1 != scanf("%f", it)) {
             printf("Error");
             return 1;
         }
     
     float avg = 0.f;
-    for (int *it = nums, *end = nums + 5; it != end; ++it)
+    for (float *it = nums, *end = nums + 5; it != end; ++it)
         avg += *it;
     avg /= 5.f;
     printf("%.3f\n", avg);
